feat: coupon distribution user services

+++ coupon-service/coupon-distribution/src/main/java/com/dongliu/coupon/repository/CouponRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public interface CouponRepository extends JpaRepository<Coupon, Integer> {
 
-    List<Coupon> findAllByUserIdAndStatus(Integer userId, CouponStatus status);
+    List<Coupon> findAllByUserIdAndStatus(Long userId, CouponStatus status);
 
     List<Coupon> findAllByUserId(Integer userId);
 }
